** added adccommands.lua  ** changes in DetectionEntry - +2 INFMaps, now there are fields for 2 protocols + for both (only one used at check)  ** some code clean-up  ** some improvements on checking external ip  ** prevent crash on ip update TODO: remove MyINFO check and add replacement partial commit, crise - do the job ;)

+++ trunk/client/CryptoManager.cpp
@@ -336,7 +336,7 @@ SSLSocket* CryptoManager::getServerSocket(bool allowUntrusted) throw(SocketExcep
 }
 
 
-void CryptoManager::decodeBZ2(const uint8_t* is, size_t sz, string& os) throw (CryptoException) {
+void CryptoManager::decodeBZ2(const uint8_t* is, unsigned int sz, string& os) throw (CryptoException) {
 	bz_stream bs = { 0 };
 
 	if(BZ2_bzDecompressInit(&bs, 0, 0) != BZ_OK)
@@ -344,7 +344,7 @@ void CryptoManager::decodeBZ2(const uint8_t* is, size_t sz, string& os) throw (C
 
 	// We assume that the files aren't compressed more than 2:1...if they are it'll work anyway,
 	// but we'll have to do multiple passes...
-	size_t bufsize = 2*sz;
+	unsigned int bufsize = 2 * sz;
 	boost::scoped_array<char> buf(new char[bufsize]);
 	
 	bs.avail_in = sz;
@@ -437,5 +437,5 @@ string CryptoManager::makeKey(const string& aLock) {
 
 /**
  * @file
- * $Id: CryptoManager.cpp 382 2008-03-09 10:40:22Z BigMuscle $
+ * $Id: CryptoManager.cpp 393 2008-06-25 18:33:20Z BigMuscle $
  */
